+++ sdk/storage/azure_storage_blob/src/clients/page_blob_client.rs
@@ -144,15 +144,14 @@ impl PageBlobClient {
     ///
     /// # Arguments
     ///
-    /// * `content_length` - The maximum size for the Page blob, up to 1TB. The page blob size must
-    ///   be aligned to a 512-byte boundary.
+    /// * `size` - The maximum size for the Page blob, up to 1TB. The page blob size must be aligned to a 512-byte boundary.
     /// * `options` - Optional configuration for the request.
     pub async fn create(
         &self,
-        content_length: u64,
+        size: u64,
         options: Option<PageBlobClientCreateOptions<'_>>,
     ) -> Result<Response<PageBlobClientCreateResult, NoFormat>> {
-        self.client.create(content_length, options).await
+        self.client.create(size, options).await
     }
 
     /// Clears a range of pages.