+++ sdk/storage/azure_storage_blob/CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Changed conversion implementation from `BlobTags` to `HashMap<String, String>` from `TryFrom` to `From`.
 - Added `continuation_token` to `PagerOptions` for methods that return a `Pager`.
+- Renamed `content_length` to `size` for `PageBlobClient`'s `create()` method.
 
 ### Breaking Changes